old-tests/submission/Infraware/Forms/contents/Forms/Not Use/input_alt.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_alt</h3>
	</p>
	
	<li>You can see [alt text].</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  id="input_form">
	<p><input type='image' id='input_image'></p>
	</form>

	<script>
		
		var input_image = document.getElementById("input_image");
		input_image.alt = 'alt text';

		if (typeof(input_image.alt) == "string") {
			if (input_image.alt != "alt text") {
				test(function() {
					assert_true(false, "alt attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("alt attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
